feat: remove module of location
