--- src/admin-location/admin-location.module.ts
@@ -1,12 +1,0 @@
-import { Module } from '@nestjs/common';
-import { AdminLocationService } from './admin-location.service';
-import { AdminLocationController } from './admin-location.controller';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { Location } from 'src/location/entities/location.entity';
-
-@Module({
-  imports: [TypeOrmModule.forFeature([Location])],
-  controllers: [AdminLocationController],
-  providers: [AdminLocationService],
-})
-export class AdminLocationModule {}
