--- src/admin-location/admin-location.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { AdminLocationService } from './admin-location.service';
-
-describe('AdminLocationService', () => {
-  let service: AdminLocationService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [AdminLocationService],
-    }).compile();
-
-    service = module.get<AdminLocationService>(AdminLocationService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
