--- src/admin-location/dto/create-admin-location.dto.ts
@@ -1,19 +1,0 @@
-import { IsNotEmpty, IsNumber, IsString } from 'class-validator';
-
-export class CreateAdminLocationDto {
-  @IsNumber()
-  @IsNotEmpty()
-  latitude: number;
-
-  @IsNumber()
-  @IsNotEmpty()
-  longitude: number;
-
-  @IsString()
-  @IsNotEmpty()
-  name: string;
-
-  @IsString()
-  @IsNotEmpty()
-  description: string;
-}
