--- src/admin-location/dto/update-admin-location.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateAdminLocationDto } from './create-admin-location.dto';
-
-export class UpdateAdminLocationDto extends PartialType(CreateAdminLocationDto) {}
