--- src/location/dto/create-location.dto.ts
@@ -1,1 +1,0 @@
-export class CreateLocationDto {}
