--- src/location/dto/update-location.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateLocationDto } from './create-location.dto';
-
-export class UpdateLocationDto extends PartialType(CreateLocationDto) {}
