--- src/location/location.controller.spec.ts
@@ -1,20 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { LocationController } from './location.controller';
-import { LocationService } from './location.service';
-
-describe('LocationController', () => {
-  let controller: LocationController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [LocationController],
-      providers: [LocationService],
-    }).compile();
-
-    controller = module.get<LocationController>(LocationController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
