+++ src/proposal/dto/create-proposal-report.dto.ts
@@ -1,5 +1,14 @@
 import { IsEnum, IsNotEmpty, IsOptional, IsString } from 'class-validator';
-import { ReportType } from 'src/user/entities/user-proposal-report.entity';
+
+export enum ReportType {
+  SPAM,
+  INAPPROPRIATE,
+  HATE_SPEECH,
+  HARASSMENT,
+  FALSE_INFORMATION,
+  PERSONAL_INFO,
+  ETC,
+}
 
 export class CreateProposalReportDto {
   @IsEnum(ReportType)
