+++ src/proposal/proposal.controller.ts
@@ -15,8 +15,6 @@ import { CreateProposalDto } from './dto/create-proposal.dto';
 import { UpdateProposalDto } from './dto/update-proposal.dto';
 import { UserId } from 'src/common/decorator/user-id.decorator';
 import { PagePaginationDto } from 'src/common/dto/page-pagination.dto';
-import { ToggleAgreementDto } from './dto/toggle-agreement.dto';
-import { CreateProposalReportDto } from './dto/create-proposal-report.dto';
 
 @Controller('proposal')
 @UseInterceptors(ClassSerializerInterceptor)
@@ -53,27 +51,5 @@ export class ProposalController {
   @Delete(':id')
   remove(@Param('id') id: number, @UserId() userId: number) {
     return this.proposalService.remove(id, userId);
-  }
-
-  @Post(':id/agreement')
-  toggleProposalAgreement(
-    @Param('id') id: number,
-    @Body() toggleAgreementDto: ToggleAgreementDto,
-    @UserId() userId: number,
-  ) {
-    return this.proposalService.toggleProposalAgreement(
-      id,
-      userId,
-      toggleAgreementDto.isAgree,
-    );
-  }
-
-  @Post(':id/report')
-  reportProposal(
-    @Param('id') id: number,
-    @Body() createReportDto: CreateProposalReportDto,
-    @UserId() userId: number,
-  ) {
-    return this.proposalService.reportProposal(id, createReportDto, userId);
   }
 }
