+++ src/proposal/proposal.module.ts
@@ -3,21 +3,10 @@ import { ProposalService } from './proposal.service';
 import { ProposalController } from './proposal.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Proposal } from './entities/proposal.entity';
-import { Location } from 'src/location/entities/location.entity';
 import { CommonModule } from 'src/common/common.module';
-import { ProposalAgreement } from './entities/proposal-agreement.entity';
-import { UserProposalReport } from 'src/user/entities/user-proposal-report.entity';
 
 @Module({
-  imports: [
-    TypeOrmModule.forFeature([
-      Proposal,
-      Location,
-      ProposalAgreement,
-      UserProposalReport,
-    ]),
-    CommonModule,
-  ],
+  imports: [TypeOrmModule.forFeature([Proposal]), CommonModule],
   controllers: [ProposalController],
   providers: [ProposalService],
 })
