+++ src/user/entities/user.entity.ts
@@ -6,14 +6,8 @@ import {
   BeforeInsert,
 } from 'typeorm';
 import { Exclude } from 'class-transformer';
-import { Location } from 'src/location/entities/location.entity';
-import { UserLocationLike } from './user-location-like.entity';
 import { Proposal } from 'src/proposal/entities/proposal.entity';
-import { UserLocationSave } from './user-location-save.entity';
-import { ProposalAgreement } from 'src/proposal/entities/proposal-agreement.entity';
-import { UserProposalReport } from './user-proposal-report.entity';
 import { Notification } from 'src/notification/entities/notification.entity';
-import { UserLocationReport } from './user-location-report.entity';
 import { v4 as uuidv4 } from 'uuid';
 import { BaseTableEntity } from 'src/common/entity/base-table.entity';
 
@@ -54,36 +48,9 @@ export class User extends BaseTableEntity {
   @Exclude({ toPlainOnly: true })
   phoneToken: string;
 
-  @OneToMany(() => Location, (location) => location.creator)
-  locations: Location[];
-
   @OneToMany(() => Proposal, (proposal) => proposal.creator)
   proposals: Proposal[];
 
   @OneToMany(() => Notification, (notification) => notification.user)
   notifications: Notification[];
-
-  @OneToMany(() => UserLocationLike, (ull) => ull.location)
-  likedLocations: UserLocationLike[];
-
-  @OneToMany(() => UserLocationSave, (save) => save.location)
-  savedLocations: UserLocationSave[];
-
-  @OneToMany(
-    () => ProposalAgreement,
-    (proposalAgreement) => proposalAgreement.proposal,
-  )
-  proposalAgreements: ProposalAgreement[];
-
-  @OneToMany(
-    () => UserProposalReport,
-    (userProposalReport) => userProposalReport.proposal,
-  )
-  proposalReports: UserProposalReport[];
-
-  @OneToMany(
-    () => UserLocationReport,
-    (userLocationReport) => userLocationReport.location,
-  )
-  locationReports: UserLocationReport;
 }
